+++ src/MartinGeorgiev/Doctrine/DBAL/Types/IntegerArray.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace MartinGeorgiev\Doctrine\DBAL\Types;
 
+use MartinGeorgiev\Doctrine\DBAL\Type;
+
 /**
  * Implementation of PostgreSQL INTEGER[] data type.
  *
@@ -17,7 +19,7 @@ class IntegerArray extends BaseIntegerArray
     /**
      * @var string
      */
-    protected const TYPE_NAME = 'integer[]';
+    protected const TYPE_NAME = Type::INTEGER_ARRAY;
 
     protected function getMinValue(): int
     {